WBUG-60 Rewrite breadcrumb system to use named routes and allow token replacement
The breadcrumb system has been rewritten now that all routes are named. This makes building the breadcrumbs easier and allows us to rely on the .NET routing system rather than trying to guess and replicate routes.

Token replacement now works which takes route values from the current route and populates breadcrumb routes with these values if a matching param name is found for the breadcrumb route. This means linking to parent pages that require a param now work.

One example of this is https://localhost:44323/projects/1/users-and-roles/2/assign
This page can now correctly show the route to the parent (users and roles) with the productId (1) populated correctly.